Issue #20: Pre-pone T_BEGIN to December 31, 1917
Prepone T_BEGIN to December 31, 1917.
*Reason*:
The GOV inaccurately dates some entries starting from 1918 which
actually only start with the end of WWI on November 11, 1918.
Example: "Volksstaat Württemberg" object_884658 only existed after the
end of WWI but in the GOV says it has the date range 1918-1945 which is
not precise enough. So this is a balance here. We might loose some rare
entries in the Verlustlisten that truly refer to changes between
January 1, 1918 and November 11, 1918 but I think these are less (if
there are any) than those objects that _inaccurately_ start from
January 1, 1918.